# Proof-of-Concept: Fast path for search